+++ Changelog.md
@@ -19,12 +19,14 @@ __New Features__
   - CEER (Combined Energy Efficiency Ratio) instead of EER for room ACs.
   - `UsageBin` instead of `FirstHourRating` (for water heaters w/ UEF metric).
 - Allows more defaulting (optional inputs):
+  - Mechanical ventilation fan airflow rate per ASHRAE 62.2-2019.
   - HVAC/DHW system efficiency (by age).
   - Mechanical ventilation fan power (by type).
   - Color (solar absorptance) for walls, roofs, and rim joists.
   - Foundation wall distance to top/bottom of insulation.
   - Door azimuth.
   - Radiant barrier grade.
+  - Whole house fan airflow rate and fan power.
 - Switches to the EnergyPlus Fan:SystemModel object for all HVAC systems.
 - Introduces a small amount of infiltration for unvented spaces.
 - Revises shared mechanical ventilation preconditioning control logic to operate less often.